photo-search/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-search"
version = "0.1.0"
description = "AI-powered personal photo search system"
requires-python = ">=3.12,<3.14"
dependencies = [
    "insightface>=0.7.3",
    "onnxruntime>=1.16.0",
    "ollama>=0.4.0",
    "qdrant-client>=1.7.0",
    "psycopg2-binary>=2.9.9",
    "Pillow>=10.0.0",
    "pillow-heif>=0.14.0",
    "reverse_geocoder>=1.5.1",
    "numpy>=1.26.0",
    "scikit-learn>=1.3.0",
    "typer[all]>=0.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "opencv-python-headless>=4.8.0",
]

[project.scripts]
photo-search = "photo_search.cli:app"

[tool.setuptools.packages.find]
include = ["photo_search*"]
